Bull game issues fixed
Fixed the issue where the game would freeze when restarting the game. Now the game will return to the start menu when the user restarts the game. I also fixed the issue where diagonal movement would not work. The issue was because the movement logic was in 'else if' statement blocks when it needed to be 'if' statements so like 'up' and 'left' could be triggered together.