Reduced motion warning on game
If the user has "reduced motion" enabled on their browser the user will be prompted a warning message that will warn them that the game has motion objects that could bother them. This focuses on the users preferences and gives a clear warning of the potential issues they might have with the game.